Fixed sub-sprite drawing and embedded shader in c code.
The sub-sprite drawing was using the size of the sprite on the screen
rather then the size of the texture when drawing, now it is using the
correct one.
Embedded the shader code into the c code, making easier to use this
library, without the need to have the shaders code in the same directory
when running a program that uses it. I'll probably make some kind of
script to generate files containing the C string with the shader code
from glsl files, because writing glsl in C string is a pain.